Add inputs and intermediate providers systems
Intermediate providers provide a mapping file of obf names to intermediate names.
Each intermediate provider has a name.

Files under the subdirectories of the input directory are parsed according to
the simple input file format (see SimpleInputFileReader), using the name of the
directory as the intermediate provider to use for parsing the files under it.

There are three intermediate providers available:

 - `obf`: obfuscated names. Always available
 - `official`: official or mojmap names. Always available
 - `srg`: SRG names, from the Forge MCPConfig project.

The artifact used by the SRG provider can be changed using the
`mcpconfig` configuration.