Use unmapped names if a name does not have a mapping entry
Intermediates (such as SRG) do not necessarily contain mappings for
all possible entries. To allow inputs to entries without
corresponding mappings, use the direct input names in those cases.

This should still be safe as the data combine method still checks for
the existence of the entry in the base data before adding.