Add new mapping IO format enum, add format properties to extension and tasks
To minimize the pain of transitioning data from one format to another, the
extensions and tasks now have a new format property for data inputs and outputs.

This format property uses the new MappingIOFormat enum, which lists out
the possible formats that may be used for reading and writing.

Specifically, this is done to prepare the way for transitioning to the
Enigma file format, the IO for which will be added by a later commit.