Build map for class metadata and use instead of manual search
Previously, the manual searching of the class in the metadata was implemented
in a way that meant that the metadata for inner classes would never be found.

Now, a map is built beforehand of class name -> class metadata, including
inner classes to fix the above issue. This also gives a performance
boost, since we do not need to iterate the class metadata set each time
to find the classes.