Add validator & sanitizer for parameter indexes
The validator checks if the parameter exists according to the owning
method's descriptor. The sanitizer removes these invalid parameters.

Currently, because of ParchmentMC/Blackstone#3, we are ignoring metadata
when calculating indexes from the descriptors. When that issue is fixed,
both validator and sanitizer should be adjusted to use it again.

Implements and closes #7