Fix class name component splitting in Enigma storage logic
Previously, because the direct output of `CharMatcher.is('$').toString()`
is passed into `String#split(String)` which takes in a regular expression,
the returned array is always one entry in size: the string itself. (Since
`$` indicates end of string, and the toString output is "$", it would try
to split on the end of the string, which means the string is unaffected).

An example of the effect of this is `RenderStateShard$MultiTextureStateShard$Builder`
being in the wrong position in the Enigma format writer's output. Instead
of being directly after the `MultiTextureStateShard` superclass, it would
be sorted to the very end of the file, which places it under the
`OffsetTexturingStateShard` superclass.

This commit fixes it by escaping the dollar sign using `Pattern#quote`,
which escapes the `$` metacharacter into "\\$", which matches against
literal `$` characters.