Fix issue with capital 'L' in reference types
A refactor from a single switch-block to multiple if-blocks is required
to fix the logic so capital 'L's in the middle of reference types (or,
as called in the comments, L-types) are not mistaken as the end of the
reference type, due to case fallthrough and no clean construct to
'skip' a case.

A regression test is also added for this issue.