Improve Enigma IO to reduce file IO churn
Previously, the Enigma data IO would delete all the files and recreate
them one-by-one, which creates a lot of churn when rewriting to the
same folder with anything less than all-the-files changes.

This new implementation avoids that by reusing mapping files if they
match a class name, deleting files which don't match a class name (or
the two JSON files) or directories without any files, and creating new
mapping files for class names which don't have a reused file.

Fixes #17