Fix accidental async download in DownloadUtil
Gradle Download Task 5.x changed DownloadAction#execute to be
asynchronous, though we assume it to be synchronous. Somehow I failed to
 see that in the changelogs.

Fixes an issue where a download would start asynchronously but a
checksum verify is done on the main thread immediately afterwards,
which fails due to the file not yet being finished downloading.