Fix task dependency inference for VersionDownload outputs
For task dependency inference to work properly for task outputs, the
instance returned by the property getter must be an instance of (a
subclass of) `Property`, even if it's only exposed as a `Provider`.

A simple workaround is to create a `Property` instance (like a
`RegularFileProperty`), and set its conventional value to what we want
it to be.