Modify CopyData to accept directories or files
Gradle doesn't have a good way to define a property input that can be
both a file or a directory. While one could use a
`Property<FileSystemLocation>` for the property type, the problem is
that both `@InputFile` and `@InputDirectory` (or corresponding output
annotations) cannot be applied to that, and neither can both be applied
to the same property.

The workaround here is to declare two separate inputs/outputs, one for
a file and one for a directory, which are both optional (so Gradle
doesn't complain when either is missing). Then, we verify at task
execution time that only one of the properties were configured.